+++ src/MQUnifiedsensor.h
@@ -274,7 +274,7 @@ class MQUnifiedsensor
     int readSensor(String nameLectureRequeired = "", bool print = false);
     int readPPM(int m, int b);
     
-    int calibrate(boolean print = false);
+    float calibrate(boolean print = false);
     double getVoltage();
     double stringToDouble(String & str);
